core/requirements.txt
asgiref==3.8.1
autopep8==2.1.0
channels==4.1.0
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gprof2dot==2022.7.29
mysqlclient==2.2.4
pillow==10.3.0
pycodestyle==2.11.1
PyJWT==2.8.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0